src/main/java/Animals.java
public class Animals {
    private int id;
    private String name;

    public Animals (int id, String name){
        this.id= id;
        this.name = name;
    }
    public int getId(){
        return id;
    }
    public String getName(){
        return name;
    }
    @Override
    public boolean equals(Object otherAnimals){
        if (!(otherAnimals instanceof Animals)) {
            return false;
        } else {
            Animals newAnimals = (Animals) otherAnimals;
            return this.getName().equals(newAnimals.getName()) &&
                    this.getName().equals(newAnimals.getName());
        }
    }
}
